infrastructure/requirements.txt
aws-cdk-lib==2.147.3
constructs==10.3.0
pytest==8.2.2
pytest-cov==5.0.0
